deleted text component
